New testing framework (#561)
* Fixed spelling

* Start on reworking the testing system

* Added basic assignment test and applied necessary changes to make it hpn

* Expanded assignment

* superusers can now see everything

* Refactored old test files

* Added basic course tests

* Added group test, and reworken group to work with rest

* Fixed bug where participation gets removed when group is deleted

* isort

* Merge branch 'develop' of github.com:eJourn-al/eJournal into epic/testing

* Setup factoryboy

* [WIP] Factory Boy

* Initial working tests with factoy boy

* pytest

* used init in factory

* Added intensive user testing

* Added user test

* Fix pytest & coverage (might not work with online yet)

* Fix directory where .coverage is omitted (now top level)

* Fix user test

* Use subfactory since one-to-one

* Removed old code and added test assignment deletion

* `make test-back test=course` now only tests course tests

* Fixed spelling

* test-back only tests for isort, and does not correct it automatically

* Improved test course

* Added group testing

* Start of format testing, first reworked it a bit

* Added basic format testing

* Switch

* isort

* Fixed journal testing

* isort

* remove print

* Added journal testing

* Added admin test

* rename: User -> Student

* Start entry testing

* Added entry required fields check backend

* Added name check

* isort

* flake

* fixed variable id

* isort

* Added flake test

* Entry tests and rework

* Alot of name changes and improved entry testing

* Type and unused parameter fix

* Added comment tests

* Fix linter issues

* Added lti tests

* Test LTI passback

* Added email tests

* Fixed KeyError responses

* Added instance tests

* Linter fix

* Added role tests

* Fixed test issues

* Added participation tests

* added htmlcov

* linter fixes

* Removed ols files

* Linter fixes

* More linter fixes

* Fixed tests

* Fixed requested changes

* ImproVED TIMELine

still needs fixing by expanding facotry boy

* linter

* fixed deprecation warnings

* removed timeline tests

* Fixed requested changes

+++ src/django/VLE/views/user.py
@@ -42,7 +42,7 @@ class UserView(viewsets.ViewSet):
         Returns:
         On failure:
             unauthorized -- when the user is not logged in
-        On succes:
+        On success:
             success -- with the course data
 
         """
@@ -99,7 +99,7 @@ class UserView(viewsets.ViewSet):
             unauthorized -- when the user is not logged in
             bad request -- when email/username/lti id already exists
             bad request -- when email/password is invalid
-        On succes:
+        On success:
             success -- with the newly created user data
         """
 
@@ -217,19 +217,19 @@ class UserView(viewsets.ViewSet):
         On success:
             success -- deleted message
         """
-        if not request.user.is_superuser:
-            return response.forbidden('You are not allowed to delete a user.')
-
         if int(pk) == 0:
             pk = request.user.id
-
         user = User.objects.get(pk=pk)
 
-        if len(User.objects.filter(is_superuser=True)) == 1:
+        if not (request.user.is_superuser or request.user == user):
+            return response.forbidden('You are not allowed to delete a user.')
+
+        # Deleting the last superuser should not be possible
+        if user.is_superuser and User.objects.filter(is_superuser=True).count() == 1:
             return response.bad_request('There is only 1 superuser left and therefore cannot be deleted')
 
         user.delete()
-        return response.deleted(description='Sucesfully deleted user.')
+        return response.success(description='Successfully deleted user.')
 
     @action(['patch'], detail=False)
     def password(self, request):
@@ -280,7 +280,7 @@ class UserView(viewsets.ViewSet):
         user = User.objects.get(pk=pk)
 
         # Check the right permissions to get this users data, either be the user of the data or be an admin.
-        if not (user.is_superuser or request.user.id == pk):
+        if not (request.user.is_superuser or request.user.id == pk):
             return response.forbidden('You are not allowed to view this user\'s data.')
 
         profile = UserSerializer(user).data
